Add files via upload
In this homework assignment, task is to program your own AI assistant that you can access via your smartphone. You will set up a Telegram bot that will serve as a communication channel to interact with a server (e.g. your laptop or PC). The server will execute the inference of an LLM (e.g. the TinyLlama model) to process and respond to your input message.